+++ Changelog.md
@@ -2,8 +2,12 @@
 
 ### Ignition Fuel Tools 4.X.X (20xx-xx-xx)
 
+1. Added `edit` subcommand to the `ign fuel`. The edit command currently
+   supports editing a model's privacy.
+    * [Pull request 67](https://github.com/ignitionrobotics/ign-fuel-tools/pull/67)
+
 1. Alphabetical listing of subcommands.
-    * [BitBucket pull request 65](https://github.com/ignitionrobotics/ign-fuel-tools/pull/65)
+    * [Pull request 65](https://github.com/ignitionrobotics/ign-fuel-tools/pull/65)
 
 ### Ignition Fuel Tools 4.1.0 (2020-02-27)